requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
feedparser==6.0.11
